Rename frontend-pipeline to frontend
